requirements.txt
#general modules
Django==3.2.25
django-filter
djangorestframework
python-decouple
requests
 
#third-party modules
djangorestframework_simplejwt
django-filter
Markdown
drf-yasg[validation]

# refactoring and tests
black
flake8
pytest
pytest-django
Faker

# background process and cache
celery
Redis
django-celery-beat